example/bottle_app/requirements.txt
bottle
python-dotenv
ikon-auth>=0.1.5
